tourist_scheduling_system/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "a2a-summit-demo"
version = "1.0.0"
description = "Multi-Agent Tourist Scheduling System with Real-time UI"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "A2A Summit Demo Team"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.8"
dependencies = [
    "a2a-sdk>=0.3.0",
    "click>=8.0.0",
    "httpx>=0.24.0",
    "uvicorn>=0.23.0",
    "fastapi>=0.100.0",
    "websockets>=11.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
a2a-scheduler = "a2a_summit_demo.agents.scheduler_agent:main"
a2a-guide = "a2a_summit_demo.agents.guide_agent:main"
a2a-tourist = "a2a_summit_demo.agents.tourist_agent:main"
a2a-ui = "a2a_summit_demo.agents.ui_agent:main"
a2a-autonomous-guide = "a2a_summit_demo.agents.autonomous_guide_agent:main"
a2a-autonomous-tourist = "a2a_summit_demo.agents.autonomous_tourist_agent:main"

[project.urls]
Homepage = "https://github.com/muscariello/a2a-summit-demo"
Repository = "https://github.com/muscariello/a2a-summit-demo"
Issues = "https://github.com/muscariello/a2a-summit-demo/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.venv
  | \.git
  | __pycache__
)/
'''

[tool.isort]
profile = "black"
src_paths = ["src", "tests"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true